remove perl from test recommendations
Newer versions of CPAN.pm install recommendations by default.  They
treat them as hard dependencies.  Since a dependency on perl is
impossible for CPAN.pm to fulfill, it will refuse to install Try::Tiny
on perl 5.8 with the default settings.

There isn't really anything useful to be done with a recommendation on a
newer perl version, so just remove it.